correct upload + tests
