+++ vip-api/src/test/java/fr/insalyon/creatis/vip/api/rest/itest/data/DataControllerIT.java
@@ -1,257 +1,86 @@
 package fr.insalyon.creatis.vip.api.rest.itest.data;
 
-import fr.insalyon.creatis.vip.api.model.PathProperties;
+import fr.insalyon.creatis.grida.common.bean.Operation;
 import fr.insalyon.creatis.vip.api.rest.config.BaseRestApiSpringIT;
-import fr.insalyon.creatis.vip.datamanager.models.PoolOperation;
-import fr.insalyon.creatis.vip.datamanager.models.PoolOperation.Status;
-import fr.insalyon.creatis.vip.datamanager.models.PoolOperation.Type;
-
+import fr.insalyon.creatis.vip.api.rest.mockconfig.DataConfigurator;
+import fr.insalyon.creatis.vip.datamanager.integrationtest.StorageTestConfigurer;
 import org.apache.commons.io.FileUtils;
 import org.hamcrest.Matchers;
-import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.io.TempDir;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mockito;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 
 import java.io.File;
 import java.nio.file.Files;
+import java.nio.file.Path;
 import java.nio.file.Paths;
 
-import static fr.insalyon.creatis.vip.api.data.PathTestUtils.*;
-import static fr.insalyon.creatis.vip.api.data.UserTestUtils.baseUser1;
+import static fr.insalyon.creatis.vip.api.data.PathTestUtils.getAbsolutePath;
+import static fr.insalyon.creatis.vip.api.data.PathTestUtils.testDir1;
 import static fr.insalyon.creatis.vip.api.data.UserTestUtils.baseUser2;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.mockito.Mockito.*;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@Disabled
-public class DataControllerIT extends BaseRestApiSpringIT {
+public class DataControllerIT extends BaseRestApiSpringIT  {
 
-    @Test
-    public void shouldReturnFilePath() throws Exception {
-        configureDataFS();
-        String testLfcPath = getAbsolutePath(testFile1);
-        mockMvc.perform(
-                get("/rest/path" + testLfcPath).param("action", "properties").with(baseUser1()))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(content().contentType(MediaType.APPLICATION_JSON_VALUE))
-                .andExpect(jsonPath("$", jsonCorrespondsToPath(testFile1PathProperties)));
+    @Autowired StorageTestConfigurer utils;
+
+    protected void configureDataFS() throws Exception {
+        DataConfigurator.configureFS(utils);
     }
 
     @Test
-    public void shouldReturnDirectoryPath() throws Exception {
-        configureDataFS();
-        String testLfcPath = getAbsolutePath(testDir1);
-        mockMvc.perform(
-                get("/rest/path" + testLfcPath).param("action", "properties").with(baseUser2()))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(content().contentType(MediaType.APPLICATION_JSON_VALUE))
-                .andExpect(jsonPath("$", jsonCorrespondsToPath(getPathWithTS(testDir1PathProperties))));
-    }
-
-    @Test
-    public void shouldReturnNonExistingPath() throws Exception {
-        String testLfcPath = "/vip/Home/WRONG/PATH";
-        when(lfcBusiness.exists(eq(baseUser1), eq(testLfcPath)))
-            .thenReturn(false);
-        PathProperties expectedPathProperties = new PathProperties();
-        expectedPathProperties.setExists(false);
-        expectedPathProperties.setPath(testLfcPath);
-        mockMvc.perform(
-                get("/rest/path" + testLfcPath).param("action", "exists").with(baseUser1()))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(content().contentType(MediaType.APPLICATION_JSON_VALUE))
-                .andExpect(jsonPath("$.exists").value(false));
-    }
-
-    @Test
-    public void shouldListuser2Dir() throws Exception {
-        configureDataFS();
-        String lfcPath = getAbsolutePath(user2Dir);
-        mockMvc.perform(
-                get("/rest/path" + lfcPath).param("action", "list").with(baseUser2()))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(content().contentType(MediaType.APPLICATION_JSON_VALUE))
-                .andExpect(jsonPath("$[*]", Matchers.containsInAnyOrder(
-                        jsonCorrespondsToPath(testDir1PathProperties),
-                        jsonCorrespondsToPath(testFile2PathProperties)
-                ))
-        );
-    }
-
-    @Test
-    public void shouldListDirectory1() throws Exception {
-        configureDataFS();
-        String lfcPath = getAbsolutePath(testDir1);
-        mockMvc.perform(
-                get("/rest/path" + lfcPath).param("action", "list").with(baseUser2()))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(content().contentType(MediaType.APPLICATION_JSON_VALUE))
-                .andExpect(jsonPath("$[*]", Matchers.containsInAnyOrder(
-                        jsonCorrespondsToPath(testFile3PathProperties),
-                        jsonCorrespondsToPath(testFile4PathProperties),
-                        jsonCorrespondsToPath(testFile5PathProperties)
-                        ))
-                );
-    }
-
-    @Test
-    public void shouldDownload() throws Exception {
-        configureDataFS();
-        String lfcPath = getAbsolutePath(testFile1);
-        String operationId = "testOpId";
-        String testFile = Paths.get(ClassLoader.getSystemResource("testFile.txt").toURI())
-                .toAbsolutePath().toString();
-        PoolOperation donePoolOperation = new PoolOperation(operationId,
-                null, null, null, testFile, Type.Download, Status.Done, baseUser1.getEmail(), 100);
-        PoolOperation runningPoolOperation = new PoolOperation(operationId,
-                null, null, null, null, Type.Download, Status.Running, baseUser1.getEmail(), 0);
-        when (transferPoolBusiness.downloadFile(eq(baseUser1), eq(lfcPath)))
-            .thenReturn(operationId);
-        when (transferPoolBusiness.getOperationById(
-                  eq(operationId), eq(baseUser1.getFolder())))
-            .thenReturn(runningPoolOperation, runningPoolOperation, donePoolOperation);
-        when (transferPoolBusiness.getDownloadPoolOperation(operationId))
-                .thenReturn(donePoolOperation);
-        mockMvc.perform(
-                get("/rest/path" + lfcPath).param("action", "content").with(baseUser1()))
-                .andDo(print())
-                .andExpect(status().isOk());
-        // todo test file content
-    }
-
-    @Test
-    public void shouldHaveDownloadTimeout() throws Exception {
-        configureDataFS();
-        String lfcPath = getAbsolutePath(testFile1);
-        String operationId = "testOpId";
-        PoolOperation runningPoolOperation = new PoolOperation(operationId,
-                null, null, null, null, Type.Download, Status.Running, baseUser1.getEmail(), 0);
-        when (transferPoolBusiness.downloadFile(eq(baseUser1), eq(lfcPath)))
-            .thenReturn(operationId);
-        when (transferPoolBusiness.getOperationById(
-                  eq(operationId), eq(baseUser1.getFolder())))
-            .thenReturn(runningPoolOperation, runningPoolOperation);
-        mockMvc.perform(
-                get("/rest/path" + lfcPath).param("action", "content").with(baseUser1()))
-                .andDo(print())
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    public void shouldDownloadAfterFirstTimeout() throws Exception {
-        configureDataFS();
-        String lfcPath = getAbsolutePath(testFile1);
-        String operationId = "testOpId";
-        String testFile = Paths.get(ClassLoader.getSystemResource("testFile.txt").toURI())
-                .toAbsolutePath().toString();
-        PoolOperation runningPoolOperation = new PoolOperation(operationId,
-                null, null, null, null, Type.Download, Status.Running, baseUser1.getEmail(), 0);
-        PoolOperation donePoolOperation = new PoolOperation(operationId,
-                null, null, null, testFile, Type.Download, Status.Done, baseUser1.getEmail(), 100);
-        when (transferPoolBusiness.downloadFile(eq(baseUser1), eq(lfcPath)))
-                .thenReturn(operationId);
-        when (transferPoolBusiness.getOperationById(
-                eq(operationId), eq(baseUser1.getFolder())))
-                .thenReturn(runningPoolOperation);
-        when (transferPoolBusiness.getDownloadPoolOperation(operationId))
-              .thenReturn(donePoolOperation);
-        mockMvc.perform(
-                get("/rest/path" + lfcPath).param("action", "content").with(baseUser1()))
-                .andDo(print())
-                .andExpect(status().isBadRequest());
-        // now do an OK download
-        Thread.sleep(3*1000);
-          when (transferPoolBusiness.getOperationById(
-                  eq(operationId), eq(baseUser1.getFolder())))
-                  .thenReturn(runningPoolOperation, runningPoolOperation, donePoolOperation);
-        mockMvc.perform(
-                get("/rest/path" + lfcPath).param("action", "content").with(baseUser1()))
-                .andDo(print())
-                .andExpect(status().isOk());
-    }
-
-    @Test
-    public void shouldUploadFile() throws Exception {
+    public void shouldUploadFile(@TempDir Path tempDir) throws Exception {
         configureDataFS();
         String path =  getAbsolutePath(testDir1) + "/uploaded.txt";
-        byte fileContent[] = Files.readAllBytes(Paths.get(
-                ClassLoader.getSystemResource("testFile.txt").toURI()));
+        utils.configureNonExistingElementForUser(baseUser2, testDir1.getName() + "/uploaded.txt");
+        byte[] fileContent = Files.readAllBytes(Paths.get(ClassLoader.getSystemResource("testFile.txt").toURI()));
         String operationId = "testOpId";
-        PoolOperation donePoolOperation = new PoolOperation(operationId,
-                null, null, null, null, Type.Upload, Status.Done, baseUser2.getEmail(), 100);
-        PoolOperation runningPoolOperation = new PoolOperation(operationId,
-                null, null, null, null, Type.Upload, Status.Running, baseUser2.getEmail(), 0);
-        when (transferPoolBusiness.uploadFile(
-                  eq(baseUser2),
-                  anyString(),
-                  eq(getAbsolutePath(testDir1))))
+        Operation donePoolOperation = new Operation(operationId, null, null, Operation.Type.Upload, baseUser2.getEmail(), null, 100);
+        donePoolOperation.setStatus(Operation.Status.Done);
+        Operation runningPoolOperation = new Operation(operationId, null, null, Operation.Type.Upload, baseUser2.getEmail(), null, 100);
+        runningPoolOperation.setStatus(Operation.Status.Running);
+
+        String remoteDirPath = utils.getRemotePathForUser(baseUser2, testDir1.getName()).toString();
+        when (gridaPoolClient.uploadFile(
+                anyString(),
+                eq(remoteDirPath),
+                eq(baseUser2.getEmail())))
                 .thenReturn(operationId);
-        when (transferPoolBusiness.getOperationById(
-                  eq(operationId), eq(baseUser2.getFolder())))
-            .thenReturn(runningPoolOperation, runningPoolOperation, donePoolOperation);
+        when (gridaPoolClient.getOperationById(eq(operationId)))
+                .thenReturn(runningPoolOperation, runningPoolOperation, donePoolOperation);
+
+        when (server.getDataManagerPath()).thenReturn(tempDir.toString());
+        when (server.getCarminApiDownloadRetryInSeconds()).thenReturn(1);
+        when (server.getCarminApiDownloadTimeoutInSeconds()).thenReturn(1000);
+
         mockMvc.perform(
-                put("/rest/path" + path)
-                        .content(fileContent).contentType(MediaType.TEXT_PLAIN)
-                        .with(baseUser2()))
+                        put("/rest/path" + path)
+                                .content(fileContent).contentType(MediaType.TEXT_PLAIN)
+                                .with(baseUser2()))
                 .andDo(print())
                 .andExpect(status().isCreated());
+
         ArgumentCaptor<String> captor = ArgumentCaptor.forClass(String.class);
-        Mockito.verify(transferPoolBusiness).uploadFile(
-            eq(baseUser2),
-            captor.capture(),
-            eq(getAbsolutePath(testDir1)));
-        String copiedFile = captor.getValue();
+        Mockito.verify(gridaPoolClient).uploadFile(
+                captor.capture(),
+                eq(remoteDirPath),
+                eq(baseUser2.getEmail()));
+        String copiedFilePath = captor.getValue();
         File expectedFile = getResourceFromClasspath("testFile.txt").getFile();
         assertThat(
-                FileUtils.contentEquals(expectedFile, new File(copiedFile)),
+                FileUtils.contentEquals(expectedFile, new File(copiedFilePath)),
                 Matchers.is(true));
-        assertThat(copiedFile, Matchers.startsWith("/tmp"));
+        Assertions.assertTrue(copiedFilePath.startsWith(tempDir.resolve("uploads").toString()));
     }
 
-    @Test
-    public void shouldUploadBase64Data() throws Exception {
-        configureDataFS();
-        String path =  getAbsolutePath(testDir1) + "/uploaded.txt";
-        String operationId = "testOpId";
-        PoolOperation donePoolOperation = new PoolOperation(operationId,
-                null, null, null, null, Type.Upload, Status.Done, baseUser2.getEmail(), 100);
-        PoolOperation runningPoolOperation = new PoolOperation(operationId,
-                null, null, null, null, Type.Upload, Status.Running, baseUser2.getEmail(), 0);
-        when (transferPoolBusiness.uploadFile(
-                  eq(baseUser2),
-                  anyString(),
-                  eq(getAbsolutePath(testDir1))))
-            .thenReturn(operationId);
-        when (transferPoolBusiness.getOperationById(
-                  eq(operationId), eq(baseUser2.getFolder())))
-            .thenReturn(runningPoolOperation, runningPoolOperation, donePoolOperation);
-        mockMvc.perform(
-                put("/rest/path" + path)
-                        .content(getResourceAsString("jsonObjects/uploadData_1.json"))
-                        .contentType("application/carmin+json")
-                        .with(baseUser2()))
-                .andDo(print())
-                .andExpect(status().isCreated());
-        ArgumentCaptor<String> captor = ArgumentCaptor.forClass(String.class);
-        Mockito.verify(transferPoolBusiness).uploadFile(
-            eq(baseUser2),
-            captor.capture(),
-            eq(getAbsolutePath(testDir1)));
-        String copiedFile = captor.getValue();
-        File expectedFile = getResourceFromClasspath("b64decoded/uploadData_1.txt").getFile();
-        assertThat(
-                FileUtils.contentEquals(expectedFile, new File(copiedFile)),
-                Matchers.is(true));
-        assertThat(copiedFile, Matchers.startsWith("/tmp"));
-    }
 }
