+++ vip-datamanagement/pom.xml
@@ -48,6 +48,19 @@
               </execution>
             </executions>
           </plugin>
+            <plugin>
+                <!-- this adds the test code as a maven artefact to use it elsewhere -->
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-jar-plugin</artifactId>
+                <version>3.0.2</version>
+                <executions>
+                    <execution>
+                        <goals>
+                            <goal>test-jar</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
         </plugins>
     </build>
 </project>
